pyproject.toml
[tool.poetry]
name = "dns-api"
version = "0.1.0"
description = ""
authors = ["Igor Santos <igr.exe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
httpx = "^0.28.1"
uvicorn = "^0.34.0"
boto3 = "^1.36.11"
python-dotenv = "^1.0.1"
fastapi-cache2 = "^0.2.2"
redis = "^5.2.1"
tinydb = "^4.8.2"
click = "^8.1.8"
pydantic = "^2.10.6"


[tool.poetry.group.dev.dependencies]
httpie = "^3.2.4"
pytest = "^8.3.4"
moto = "^5.0.28"
ipdb = "^0.13.13"
ipython = "^8.32.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
